Reorganize modules under catalog with better domain naming
- Rename 'models' to 'items' and move under catalog/items/
- Move schema/ under catalog/schema/
- Update all imports to use catalog::items:: and catalog::schema::
- Update README.md with new hierarchical structure

This creates a cohesive catalog module containing all recipe-related
data structures, schemas, loading, and discovery logic. The 'items'
name better represents catalog contents than generic 'models'.
