+++ src/catalog/items/weighted_ingredient.rs
