+++ src/catalog/mod.rs
@@ -1,5 +1,7 @@
 mod discovery;
+pub mod items;
 mod loader;
+pub mod schema;
 
-pub use discovery::find_workspace;
+pub use discovery::{find_workspace, is_workspace};
 pub use loader::load_recipes;
