+++ src/catalog/schema/ingredients.schema.json
