+++ src/main.rs
@@ -1,8 +1,6 @@
 pub mod catalog;
 pub mod commands;
 pub mod error;
-pub mod models;
-pub mod schema;
 pub mod utils;
 
 use clap::{Parser, Subcommand};
